+++ mayan/apps/document_indexing/links.py
@@ -2,7 +2,6 @@ from __future__ import unicode_literals
 
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.documents.permissions import permission_document_type_edit
 from mayan.apps.navigation.classes import Link
 from mayan.apps.navigation.utils import get_cascade_condition
 
@@ -24,10 +23,17 @@ link_document_index_instance_list = Link(
     text=_('Indexes'), view='indexing:document_index_list',
 )
 
+link_document_template_sandbox = Link(
+    args='resolved_object.pk',
+    icon_class_path='mayan.apps.document_indexing.icons.icon_document_template_sandbox',
+    permissions=(permission_document_indexing_edit,),
+    text=_('Sandbox'),
+    view='indexing:document_template_sandbox',
+)
 link_document_type_index_templates = Link(
     args='resolved_object.pk',
     icon_class_path='mayan.apps.document_indexing.icons.icon_document_type_index_templates',
-    permissions=(permission_document_type_edit,), text=_('Index templates'),
+    permissions=(permission_document_indexing_create,), text=_('Index templates'),
     view='indexing:document_type_index_templates',
 )