+++ mayan/apps/document_indexing/urls.py
@@ -8,10 +8,11 @@ from .api_views import (
     APIIndexTemplateView, APIIndexView
 )
 from .views import (
-    DocumentIndexNodeListView, DocumentTypeIndexesView, IndexInstanceNodeView,
-    IndexListView, IndexesRebuildView, SetupIndexDocumentTypesView,
-    SetupIndexCreateView, SetupIndexDeleteView, SetupIndexEditView,
-    SetupIndexListView, SetupIndexRebuildView, SetupIndexTreeTemplateListView,
+    DocumentIndexNodeListView, DocumentTemplateSandboxView,
+    DocumentTypeIndexesView, IndexInstanceNodeView, IndexListView,
+    IndexesRebuildView, SetupIndexDocumentTypesView, SetupIndexCreateView,
+    SetupIndexDeleteView, SetupIndexEditView, SetupIndexListView,
+    SetupIndexRebuildView, SetupIndexTreeTemplateListView,
     TemplateNodeCreateView, TemplateNodeDeleteView, TemplateNodeEditView
 )
 
@@ -82,6 +83,11 @@ urlpatterns_tools = [
     url(
         regex=r'^instances/rebuild/$', view=IndexesRebuildView.as_view(),
         name='rebuild_index_instances'
+    ),
+    url(
+        regex=r'^documents/(?P<pk>\d+)/sandbox/$',
+        view=DocumentTemplateSandboxView.as_view(),
+        name='document_template_sandbox'
     ),
 ]